+++ test/controllers/internal_controller_test.rb
@@ -27,4 +27,34 @@ class InternalControllerTest < ActionController::TestCase
     get :directory
     assert_redirected_to new_user_session_path
   end
+
+  test 'should get category as viewer' do
+    login(@viewer)
+    get :category, category: categories(:one).slug
+    assert_not_nil assigns(:category)
+    assert_response :success
+  end
+
+  test 'should not get category as anonymous user' do
+    get :category, category: categories(:one)
+    assert_nil assigns(:category)
+    assert_redirected_to new_user_session_path
+  end
+
+  test 'should download document as viewer' do
+    login(@viewer)
+    get :document, category: categories(:one).slug, document_id: documents(:two)
+    assert_not_nil assigns(:category)
+    assert_not_nil assigns(:document)
+    assert_kind_of String, response.body
+    assert_equal File.binread(File.join(CarrierWave::Uploader::Base.root, assigns(:document).document.url)), response.body
+    assert_response :success
+  end
+
+  test 'should not download document as anonymous user' do
+    get :document, category: categories(:one).slug, document_id: documents(:two)
+    assert_nil assigns(:category)
+    assert_nil assigns(:document)
+    assert_redirected_to new_user_session_path
+  end
 end